client_neat/requirements.txt
neat-python
numpy
websockets
